requirements.txt
sqlalchemy
fastapi
asyncpg
uvicorn
pendulum==2.0.3
greenlet==3.0.3
cryptography==42.0.5
python-dotenv
enum-tools
bcrypt
pyjwt
asyncio
fastapi-cache2
PyJWT
pydantic[email]
alembic

